test: separate mock request into its own file
